Add photo resource methods

+++ src/app/dao/PhotoDAO.js
@@ -2,23 +2,23 @@ const pool = require('../database/Pool');
 
 module.exports = {
   
-  getAll() {
-    return pool.query('SELECT * FROM photo');
+  getUserPhotos(user_id) {
+    return pool.query('SELECT * FROM photo WHERE user_id = ?', user_id);
   },
 
-  get(id) {
-    return pool.query('SELECT * FROM photo WHERE id = ?', id);
+  getUserPhoto(id, user_id) {
+    return pool.query('SELECT * FROM photo WHERE id = ? AND user_id = ?', [id, user_id]);
   },
 
-  store(photo) {
+  createPhoto(photo) {
     return pool.query('INSERT INTO photo SET ?', photo);
   },
 
-  update(id, photo) {
-    return pool.query('UPDATE photo SET ? WHERE id = ?', [photo, id]);
+  updatePhotoDescription(description, id, user_id) {
+    return pool.query('UPDATE photo SET description = ? WHERE id = ? AND user_id', [description, id, user_id]);
   },
 
-  delete(id) {
-    return pool.query('DELETE FROM photo WHERE id = ?', id);
+  deletePhoto(id, user_id) {
+    return pool.query('DELETE FROM photo WHERE id = ? AND user_id = ?', [id, user_id]);
   }
 };
